+++ docs/source/Changelog.rst
@@ -56,6 +56,10 @@ Fixed
    -  When using a –variant and an specific target, it was generated for
       the first variant in the list, not the rest
 
+-  Gerber Drill:
+
+   -  Avoid showing an error when KiCad 10 skips a drill file (See #928)
+
 -  KiRi:
 
    -  ``kiri_server`` path to the HTML